save: guard copies with file existence checks

+++ harry-save.bash
@@ -1,12 +1,12 @@
 #!/bin/bash
-cp ~/.vimrc config/vimrc
-cp ~/.tmux.conf config/tmux.conf
+[ -f ~/.vimrc ] && cp ~/.vimrc config/vimrc
+[ -f ~/.tmux.conf ] && cp ~/.tmux.conf config/tmux.conf
 
 # save platform-specific harry bash
 case "$(uname -s)" in
-    Linux*)  cp ~/.harry.bash config/harry-linux.bash ;;
-    Darwin*) cp ~/.harry.bash config/harry-mac.bash ;;
+    Linux*)  [ -f ~/.harry.bash ] && cp ~/.harry.bash config/harry-linux.bash ;;
+    Darwin*) [ -f ~/.harry.bash ] && cp ~/.harry.bash config/harry-mac.bash ;;
 esac
 
 # saving nvim/lazyvim config:
-cp -r ~/.config/nvim/ config/nvim/
+[ -d ~/.config/nvim ] && cp -r ~/.config/nvim/ config/nvim/
